+++ Save/MainNavigationController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import TorManager
 
 class MainNavigationController: UINavigationController {
 
@@ -18,7 +19,11 @@ class MainNavigationController: UINavigationController {
 
     func setRoot() {
         if Settings.firstRunDone {
-            if !(topViewController is MainViewController) {
+            if Settings.useTor && !TorManager.shared.connected {
+                setViewControllers([UIStoryboard.main.instantiate(TorStartViewController.self)],
+                                   animated: true)
+            }
+            else if !(topViewController is MainViewController) {
                 setViewControllers([UIStoryboard.main.instantiate(MainViewController.self)],
                                    animated: true)
             }