+++ Save/Management/UploadManager.swift
@@ -13,6 +13,7 @@ import Regex
 import CleanInsightsSDK
 import BackgroundTasks
 import Photos
+import TorManager
 
 extension Notification.Name {
     static let uploadManagerPause = Notification.Name("uploadManagerPause")
@@ -61,7 +62,8 @@ class UploadManager: NSObject, URLSessionTaskDelegate {
     var waiting: Bool {
         globalPause || 
         (reachability?.connection ?? Reachability.Connection.unavailable == .unavailable) ||
-        (Settings.useOrbot && OrbotManager.shared.status == .stopped)
+        (Settings.useOrbot && OrbotManager.shared.status == .stopped) ||
+        (Settings.useTor && !TorManager.shared.connected)
     }
 
     private var current: Upload?
@@ -105,7 +107,9 @@ class UploadManager: NSObject, URLSessionTaskDelegate {
             conf.isDiscretionary = false
             conf.shouldUseExtendedBackgroundIdleMode = true
 
-            _backgroundSession = URLSession.withImprovedConf(configuration: conf, delegate: self)
+            _backgroundSession = URLSession(
+                configuration: Self.improvedSessionConf(conf),
+                delegate: self, delegateQueue: nil)
         }
 
         return _backgroundSession!
@@ -117,10 +121,23 @@ class UploadManager: NSObject, URLSessionTaskDelegate {
     */
     private var foregroundSession: URLSession {
         if _foregroundSession == nil {
-            _foregroundSession = URLSession.withImprovedConf(delegate: self)
+            _foregroundSession = URLSession(
+                configuration: Self.improvedSessionConf(),
+                delegate: self, delegateQueue: nil)
         }
 
         return _foregroundSession!
+    }
+
+
+    public class func improvedSessionConf(_ conf: URLSessionConfiguration? = nil) -> URLSessionConfiguration {
+        let conf = URLSessionConfiguration.improved(conf)
+
+        if Settings.useTor {
+            conf.connectionProxyDictionary = TorManager.shared.torSocks5ProxyConf
+        }
+
+        return conf
     }
 
 
@@ -136,6 +153,11 @@ class UploadManager: NSObject, URLSessionTaskDelegate {
         else {
             restart()
         }
+    }
+
+    func reinitSession() {
+        _backgroundSession = nil
+        _foregroundSession = nil
     }
 
     /**
@@ -540,6 +562,12 @@ class UploadManager: NSObject, URLSessionTaskDelegate {
 
             if Settings.useOrbot && OrbotManager.shared.status == .stopped {
                 self.debug("#uploadNext should use Orbot, but Orbot not started")
+
+                return self.endBackgroundTask(.noData)
+            }
+
+            if Settings.useTor && !TorManager.shared.connected {
+                self.debug("#uploadNext should use built-in Tor, but Tor not started")
 
                 return self.endBackgroundTask(.noData)
             }