+++ Shared/Projects/BrowseViewController.swift
@@ -150,33 +150,32 @@ class BrowseViewController: BaseTableViewController {
         beginWork {
             data.removeAll()
 
-            if let space = SelectedSpace.space as? WebDavSpace, let url = space.url {
-                space.session.info(url) { info, error in
-                    if let error = error {
-                        self.data.append([error])
-
-                        return
-                    }
-
-                    var folders = [Folder]()
-
-                    let files = info.dropFirst().sorted(by: {
-                        $0.modifiedDate ?? $0.creationDate ?? Date(timeIntervalSince1970: 0)
-                        > $1.modifiedDate ?? $1.creationDate ?? Date(timeIntervalSince1970: 0)
-                    })
-
-                    for file in files {
-                        if file.type == .directory && !file.isHidden {
-                            folders.append(Folder(file))
-                        }
-                    }
-
-                    self.data.append(folders)
-
-                    self.endWork()
-                }
+            guard let space = SelectedSpace.space as? WebDavSpace, let url = space.url else {
+                return self.endWork()
             }
-            else {
+
+            URLSession(configuration: .improved()).info(url, credential: space.credential) { info, error in
+                if let error = error {
+                    self.data.append([error])
+
+                    return
+                }
+
+                var folders = [Folder]()
+
+                let files = info.dropFirst().sorted(by: {
+                    $0.modifiedDate ?? $0.creationDate ?? Date(timeIntervalSince1970: 0)
+                    > $1.modifiedDate ?? $1.creationDate ?? Date(timeIntervalSince1970: 0)
+                })
+
+                for file in files {
+                    if file.type == .directory && !file.isHidden {
+                        folders.append(Folder(file))
+                    }
+                }
+
+                self.data.append(folders)
+
                 self.endWork()
             }
         }